LayoutTests/fast/workers/worker-navigator.html

<body>
<p>Test Navigator properties.</p>
<div id=result></div>
<script>
function log(message)
{
    document.getElementById("result").innerHTML += message + "<br>";
}

if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.waitUntilDone();
}

var worker = new Worker('resources/worker-common.js');
worker.postMessage("eval typeof navigator");
worker.postMessage("eval navigator");
worker.postMessage("eval navigator.appName");
worker.postMessage("eval navigator.appVersion.indexOf('WebKit') != 0");
worker.postMessage("eval navigator.platform.length > 0");
worker.postMessage("eval navigator.userAgent.indexOf('WebKit') != 0");
worker.postMessage("eval typeof navigator.onLine");
worker.postMessage("eval foo//bar");

worker.onmessage = function(evt) {
    if (!/foo\/\/bar/.test(evt.data))
        log(evt.data.replace(new RegExp("/.*LayoutTests"), "<...>"));
    else {
        log("DONE");
        if (window.layoutTestController)
            layoutTestController.notifyDone();
    }
}
</script>
</body>
</html>
