LayoutTests/fast/workers/worker-navigator-expected.txt

Test Navigator properties.

typeof navigator: object
navigator: [object WorkerNavigator]
navigator.appName: Netscape
navigator.appVersion.indexOf('WebKit') != 0: true
navigator.platform.length > 0: true
navigator.userAgent.indexOf('WebKit') != 0: true
typeof navigator.onLine: boolean
DONE
